Delete config-files.ts
